refactor GoalFormScreen
